requirements.txt
Flask
Flask-Mail
python-dotenv
pandas
openpyxl